+++ internal/gpscmd/gpscmd.go
@@ -77,7 +77,7 @@ func Cmd(logWriter io.Writer, logLevel slog.Level, progName string, cmdName stri
 	}
 	ctx := context.Background()
 	ctx, _ = cmd.CancelOnSignal(ctx, lg)
-	err = run(ctx, lg, target, raw, conn, v.vendor, v.packetLogPath, v.packetLogMode, v.capture, v.showReceiver, v.configSupport, args)
+	err = run(ctx, lg, target, raw, conn, v.vendor, v.packetLogPath, v.packetLogMode, v.capture, v.showReceiver, v.jsonOut, v.configSupport, args)
 	return
 }
 
@@ -151,7 +151,7 @@ func configTargetIsProbeOnly(target *gpsprot.ConfigTarget) bool {
 // Parameter dependencies:
 //   - logMode: must not be testLogMode when raw is non-nil
 //   - args: only used for test log header when logMode is testLogMode
-func run(ctx context.Context, lg *slog.Logger, target *gpsprot.ConfigTarget, raw []msgfile.RawMsg, conn gpsio.Conn, vendor gpsreg.Vendor, logPath string, logMode packetLogMode, capture opt.Val[time.Duration], showReceiver bool, support configSupportReq, args []string) error {
+func run(ctx context.Context, lg *slog.Logger, target *gpsprot.ConfigTarget, raw []msgfile.RawMsg, conn gpsio.Conn, vendor gpsreg.Vendor, logPath string, logMode packetLogMode, capture opt.Val[time.Duration], showReceiver bool, jsonOut bool, support configSupportReq, args []string) error {
 	defer func() {
 		addr := conn.LocalAddr()
 		lg.Debug("closing the GPS connection", "addr", addr)
@@ -190,7 +190,7 @@ func run(ctx context.Context, lg *slog.Logger, target *gpsprot.ConfigTarget, raw
 	if raw != nil {
 		err = runMsgs(ctx, lg, conn, pCh, raw, capture)
 	} else if target != nil {
-		rslt, err = runConfig(ctx, lg, target, pCh, conn, vendor, capture, showReceiver, support)
+		rslt, err = runConfig(ctx, lg, target, pCh, conn, vendor, capture, showReceiver, jsonOut, support)
 	} else {
 		// Passive capture mode: just read and log packets
 		if capture.IsSet() {
@@ -212,7 +212,7 @@ func run(ctx context.Context, lg *slog.Logger, target *gpsprot.ConfigTarget, raw
 	return err
 }
 
-func runConfig(ctx context.Context, lg *slog.Logger, target *gpsprot.ConfigTarget, pCh <-chan scan.Packet, conn gpsio.Conn, vendor gpsreg.Vendor, capture opt.Val[time.Duration], showReceiver bool, support configSupportReq) (*gpscfg.Result, error) {
+func runConfig(ctx context.Context, lg *slog.Logger, target *gpsprot.ConfigTarget, pCh <-chan scan.Packet, conn gpsio.Conn, vendor gpsreg.Vendor, capture opt.Val[time.Duration], showReceiver bool, jsonOut bool, support configSupportReq) (*gpscfg.Result, error) {
 	// Compile-time check: serial faults surfaced by gpsio satisfy the
 	// gpscfg.SerialError interface. gpscfg relies on this.
 	var _ gpscfg.SerialError = (*gpsio.SerialError)(nil)
@@ -223,14 +223,18 @@ func runConfig(ctx context.Context, lg *slog.Logger, target *gpsprot.ConfigTarge
 		err = nil
 	}
 	if err == nil && rslt != nil {
+		warnMissingConfigSupport(lg, support, rslt.ConfigSupport)
+		if !configTargetIsProbeOnly(target) {
+			logFailedProps(lg, &target.Props, rslt.ConfigProps)
+		}
+	}
+	if jsonOut {
+		printJSON(os.Stdout, rslt, err)
+	} else if err == nil && rslt != nil {
 		if showReceiver {
 			printReceiverInfo(os.Stdout, rslt.ReceiverInfo)
 			printConfigSupport(os.Stdout, rslt.ConfigSupport)
 			printPacketFormats(os.Stdout, rslt.PacketFormatsDetected)
-		}
-		warnMissingConfigSupport(lg, support, rslt.ConfigSupport)
-		if !configTargetIsProbeOnly(target) {
-			logFailedProps(lg, &target.Props, rslt.ConfigProps)
 		}
 		printProps(os.Stdout, rslt.ConfigProps)
 	}